[bt][hfp-hf] Query current calls procedure
Adds an implementation for the Query List of Current Calls (AT+CLCC)
procedure.  On a new call, queries the peer's call list to get direction
and maybe number of the new call. Parses the response and sets the
values for the existing calls, logging errors if these values differ
from our version of them.

Test: fx test bt-hfp-hands-free-tests
Bug: b/432243006
Change-Id: I2988c05b4d8f8912c3702ec851985871c3c3e3d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1327292
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>