[bt][hfp] Generate a NextCall even if we don't have a phone number
This is a hack.  If the +CLCC response for a given call doesn't contain
a phone number, we assume we won't get it via some other mechanism as
the peer doesn't know it.  In that case we set the phone number for the
call to the empty string. This will allow us to generate a NextCall
FIDL response for that call once we have all the other info for it.

Test: fx test bt-hfp-hands-free-tests
Bug: b/421998446
Change-Id: If220d13b065067d1fc65f260f843f93b8a38c0bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1330467
Commit-Queue: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>