[fastboot] Shred userdata on `fastboot erase userdata`
To support the `-w` flag, we need to handle the `fastboot erase` command
which is issued with the userdata, cache, and metadata partitions by the
host tool. On Fuchsia devices, we store cache inside of the data volume,
so we use the erase command for "userdata" to handle shredding the data
volume. This ensures that the data volume will be wiped on the next
boot, since we will fail to unlock it and reprovision the data volume.

We may need to revisit this approach in the future depending on how we
want to handle the metadata partition, but for now we emit a warning
when these partitions are erased, and allow the commands to succeed.

Output from a Fuchsia device:
```
~$ fastboot -w
Erasing 'userdata'                                 OKAY [  0.129s]
Erase successful, but not automatically formatting.
File system type fxfs-vol not supported.
wipe task partition not found: cache
wipe task partition not found: metadata
Finished. Total time: 0.144s
~$ echo $?
0
```

Bug: 397515768
Test: Ran on device, see output above.
Change-Id: I06748ab1ad760c2152a31878db3562bf8363aa0a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1380062
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>