[sdk][doc] Generate documentation for specific `zx::Status` values
While it would be ideal to generate all this from errors.fidl, we don't
do that today. It is useful to have this fundamental documentation,
since we need this all the time. This is the next best thing I think.

This change was generated by an automated coding assistant, with human
supervision.

Prompt:
```

Read the file `//zircon/system/public/zircon/errors`.

This file contains the documentation for Zircon errors , which are constants
which are all named starting with `ZX_ERR_`.

The file `//sdk/rust/zx-types/src/lib.rs` has a redefinition of these constants,
but in rust.

The file `//sdk/rust/zx-status/src/lib.rs` has a yet another redefinition of these
constants in rust

For both files, take the documentation for each error from the first file, and
add as rust doc comments to matching values in the other two files.

Create and upload a change based on instructions on creating changes for Fuchsia.

* For the above three files, at the beginning of the added section add a comment:
  `// LINT.IfChange(zx_status_t)`, and at end add `//LINT.ThenChange(//zircon/vdso/errors.fidl)`

* Make a corresponding change in `//zircon/vdso/errors.fidl`.

```

Change-Id: Ibedc31178b68579b8dc714332cce42f4163f9e03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1382008
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>