[usb][dwc2] Deliberately leak pinned endpoint memory.
Do the minimum possible to prevent runaway DWC2 DMA from trashing
physical memory after shutdown.

Currently, the pinned memory used by endpoints is owned by a server
implemented by the `usb-endpoint` library.  If the channel connection
to this server is closed at any point in time, the library (by
default) will Unpin the memory.

This is not safe, because no attempt is made to stop the hardware
which may be using this memory.  The proper thing to do would be to
make certain that the DMA is 100% for sure stopped before unpinning
the memory.

Unfortunately, this is harder than it looks.  Stopping individual
endpoints in the DWC2 hardware is something which seems to be only
_reluctantly_ supported.  It should be possible, but it involves an
intricate reset sequence which depends on hardware revisions.  These
sequences are further complicated by the locking situation in the
driver right now.

The driver has two types of locks: the main lock and one lock
for each endpoint.  According to comments in the driver, the endpoint
locks must always be held before the main lock.  The sequence,
however, involved manipulating a number of registers which have RWM
semantics, and need to be accessed exclusively during the reset, in
particular GRSTCTL, and various interrupt status registers (GINTSTS,
DOEPINT, DIEPINT).

In theory, all of these registers should be protected by the main
lock, but attempting to do this consistently produces a situation
where the main lock needs to be held to manipulate registers and
figure out what needs to be done (for example, in the IRQ handler),
but then requires holding an endpoint lock in order to change the
state of the endpoint in response to the interrupt.

Attempts to restructure the driver to obey some set of rules on this
front _sort-of_ work, but also seem able to producing deadlocks and
livelocks in other portions of driver host responsible for the overall
CDC ethernet link.

This is all going to need some serious attention eventually, but until
the resources can be found to clean this all up and make it safer, do
the following.

1) Correct the reset sequence and centralize it in one function.  In
   the past, the reset sequence was waiting for AHBIdle _before_
   sending the CSftRst command to the reset register.  The docs are a
   bit vague about the ordering required here, but the Synopsis
   supplied linux driver for DWC2 does the wait after the soft reset
   command has been completed, so we're going to go with that for now.
2) Require that the global lock be held during reset, preventing an
   accidental clearing of the CSftRst bit while a reset is happening
   (which could cause reset to be declared finished before it actually
   is)
3) Manually ensure that the lock is always held any time someone is
   using the reset register.  When this code is eventually cleaned up,
   this approach should be refactored to ensure that all groups of
   registers which require exclusive access are properly protected by
   a lock (or locks) in a fashion which is statically annotated to
   prevent accidents.
4) During early Init, make sure that the controller has been
   explicitly reset, and then (and only then) release the quarantine
   on the driver's BTI.
5) Also make certain that the hardware has been explicitly placed into
   reset during PrepareStop.
6) In any place where reset is called without a properly plumbed error
   handling path, terminate the driver if reset fails.  It is better
   to leak pinned memory to the BTI quarantine than run the risk of
   runaway DMA.
7) Finally, when the channel to an endpoint is closed, explicitly leak
   the pinned memory to the quarantine instead of attempting to
   recover it.

With these minimal changes in place, access to the reset register
should be safe, and the driver will be explicitly resetting itself
both during initialization, and the shutdown paths.  If any of these
resets fail, the driver will self terminate ensuring that pinned
memory should be leaked.  Finally, pinned endpoint memory will always
be leaked, however it will be recovered if the driver is simply
restarted, meaning that the leaks are not as permanent as they might
seem.

Eventually, when the driver is cleaned up, one of two things needs to
happen.

1) Endpoints will need to properly stop themselves dynamically when
   their server channels are closed before unpinning the memory.
2) The individual endpoint are eliminated completely from the design.
   Instead, the driver would allow endpoints to be allocated and
   started as needed by the driver's client, however stopping
   individual endpoints would be disallowed, allowing shutdown or
   reset of the driver to involve a simple core reset followed by
   unpinning of old memory.

Bug: 445524948
Change-Id: I3b26c82ebf694aa703b0239944fdf58efd935eeb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1382006
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Sam Hansen <hansens@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>