[kernel][geni] Implement a minimal suspend/resume sequence for Geni
Implement a minimal suspend/resume sequence for the Geni kernel UART
driver.  This sequence mostly follows the guidelines given in the
QUPv3 HW User's Guide documentation.  Specifically it:

1) Performs a graceful command cancel on the TX pipeline to ensure
   that there are no TX commands in flight when the clock is lost.
2) Performs a graceful command cancel on the RX pipeline.  This is
   considered to be "optional" by documentation, but in practice,
   failure to do this results in the RX pipeline not working after
   resuming.
3) Re-issue an RX-Go command during resume in order to restart the RX
   pipeline.

Note that this approach results in the contents of the FIFO at suspend
time being discarded.  While not ideal, it is significantly better
than simply locking up.  Follow-up changes will implement an approach
which keeps a copy of the FIFO contents in memory, and pays attention
to the FIFO status at the time of command cancel, allowing the
un-transmitted characters to be supplied to the UART again after
resume, so that they don't get dropped.

Additionally, there is no explicit control of the QUPv3 clocks going
on here (yet).  Moving forward, the hope is to shutdown all of the
clocks in the proper sequence (and restore them after waking) as part
of this operation, however there is still a bit of research which
needs to be done before this can happen.

Bug: 439065000
Change-Id: Id91508d331f3a89d1ad6b18a2909580c95578645
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1362597
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>