[bt][hfp] Delay start after stop on codec
When stopping based on an Audio request, we can receive a start very
shortly after.  This means we could still be disconnecting the current
SCO connection.  Add a delay to avoid this until we have an API signal fix.

Also, it was possible for a peer to change their Codec available between
disconnecting and reocnnecting.  This would mean we reject a new SCO
connection. Forget the CodecId when we are disconnecting.

Bug: fxbug.dev/447629864
Test: added new unit tests; fx test bt-hfp-tests
Change-Id: I44df6a56bc3382e63a5ee31e1a344b9297e493bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1381843
Reviewed-by: Matthew Kehrt <kehrt@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>