[metrics] Move Unit, TestCaseResult out of trace_processing
Unit and TestCaseResult implement the fuchsiaperf schema, which can be
(and is) used to report all manner of metrics -- both trace-derived and
not. As such, they should not live in the `trace_processing` module.
Move them to the new `reporting` module and update references.

Leave aliases behind temporarily to ease migration.

Bug: 448472515
Test: fx test trace_metrics_test trace_utils_test power_metrics_test
Change-Id: I90a2c5b9131dd0904e4ccd82085d90b4594a82c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1381188
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>