[fidl][rust-next] Correct empty-struct handling
Empty structs are represented as a single zero octet in the wire format
for C/C++ memory layout reasons. This updates the rust-next bindings and
generator to follow this and enables conformance tests that were
disabled because of this bug.

Change-Id: I3e181573c4d938aa790bf2b1916ccb09a43fe6f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1380925
Reviewed-by: David Koloski <dkoloski@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>