[starnix][timer] Implement TFD_TIMER_CANCEL_ON_SET
https://man7.org/linux/man-pages/man2/timerfd_create.2.html:

```
This flag allows monitoring for UTC timeline changes.

TFD_TIMER_CANCEL_ON_SET
      If this flag is specified along with TFD_TIMER_ABSTIME and
      the clock for this timer is CLOCK_REALTIME or
      CLOCK_REALTIME_ALARM, then mark this timer as cancelable if
      the real-time clock undergoes a discontinuous change
      (settimeofday(2), clock_settime(2), or similar).  When such
      changes occur, a current or future read(2) from the file
      descriptor will fail with the error ECANCELED.
```

Interpreting the spec, a timer that is armed with a zero realtime
absolute deadline, and configured with `TFD_TIMER_CANCEL_ON_SET`
will report that data is available when used in `epoll`, but when
read will return `ECANCELED`, not valid data. This is a special
behavior which is used to track the UTC timeline changes.

In contrast, "regular" realtime timers react to UTC timeline changes
either by expiring earlier than expected if UTC timeline has been pulled
in, or simply waiting longer than expected if UTC timeline had been
pushed out, so they do not need special treatment.

Most of the implementation effort was in bringing in a test fixture that
vends a UTC clock handle and ensures that the `settimeofday` syscall
(not the libc function!) can be called. The new syscall test had to be
disabled in the host configuration, because the host will return EPERM
on this call. This is expected but also means we can not run the test on
host. The test was moved into its own file to ensure other tests remain
operating on an immutable UTC timeline.

Another realization is the need to allow for wait_async predicates which
can be reconfigured after they are set up.

Bug: 445263761
Change-Id: I486fc43597ad8b65222192a0e7825230ec90a1b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370202
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>