[fzl] VmarManager allow using zx::unowned_vmar
Allows for creating VmarManager from a zx::unowned_vmar directly,
rather than having to use another VmarManager.

This will allow drivers to use the VmarManager to provide to the
VmoMapper so they can map into the vmar given to them by
the driver framework. This is equivalent to the root vmar as it is
managed external to the user, so it does not need to be wrapped in a
VmarManager to handle its lifetime.

Change-Id: I9ac311773314d5f92ab5a612f8d6582af2d7e3bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1379860
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>