Reland "[ffx] Refactor behavior into target/behavior:lib"
This is a reland of commit 1ca97241219f948469bdbc4136c1dbcb7f902ef3

The problem was that this was submitted after an incompatible CL
was submitted, with changes to `ffx target ssh` (fxr/1381747).
This CL now incorporates the changes port that code as well.
(Reviewing this CL can be limited to the code in target/ssh/...)

Original change's description:
> [ffx] Refactor behavior into target/behavior:lib
>
> As a precursor to changing the logic for initializing the
> FhoConnectionBehavior, we need to move the behavior logic
> into its own crate.
>
> Change-Id: I6a6a696481cfa20b7fc650b5e62c6c7aa3af7998
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1378521
> Reviewed-by: Colin Nelson <colnnelson@google.com>
> Reviewed-by: Nikita Jindal <nikitajindal@google.com>
> Commit-Queue: Steven Grady <slgrady@google.com>

Change-Id: I2665a664fe42f70d5772f6402803e5f27516ce46
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1381936
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>