[ffx] Refactor behavior initialization
Instead of having each caller set the Fho behavior, the logic for
setting it goes into the FhoTargetEnvironment. This simplifies the
callers, but more importantly it enables the upcoming change in which
the FhoTargetEnvironment ensures that no more than one behavior
object is initialized at the same time, something that can't be done if
the callers initialize the behaviors themselves.

Change-Id: I6a6a6964b8cea5e9e5306dd52ac4305273302b1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1379875
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>