[fuchsia-controller] Encapsulate all connect logic in reattempt loop.
In the event that the remote control proxy disconnects before connecting
to a specific component, it is not possible to reattempt the whole
connection again. We now pass a closure to the connection function so
that the resulting use of the RCS proxy can be reattempted in addition
to the original RCS connect/helper code.

Test: manually verified working. Testing infra in CQ exercises this codepath heavily.
Bug: 446880959
Change-Id: Ia02f9067e8a3a0c701d305aaf81ae81b6a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1378381
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>