[netstack3] Generate Random TCP Timestamp Offsets
RFC 7323 recommends adding random per connection offsets to the values
sent in the TCP timestamp option. This avoids leaking details about the
systems internal clock to the external world. The algorithm
recommended for the generation of these offsets is the same as is used
for the generation of Initial Sequence Numbers.

Bug: 360401604
Change-Id: Ie472ad1d4785688ffd1d5555d3e79e7b3f273e6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1383501
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>