[starnix][waiter] Don't use EDGE_TRIGGERED option in Zircon async waits
This removes the EDGE_TRIGGERED flag from the async wait registration
for Zircon backed objects. This is not expected to change behavior.

Starnix internally uses an event-driven model for alerting waiters.
Zircon uses a level based model for all waits. When translating a
Starnix wait to a Zircon asynchronous wait, Starnix was using the edge
triggered option to zx_object_wait_async which asks Zircon to trigger
the wait only on a rising edge on the associated signals. This is not
the same as an event-driven trigger as events are independent of the
initial level state.

When Starnix performs a blocking op on a Zircon-backed object it uses
the following sequence of operations:

1. Attempt the operation. If it says it would block, continue.
2. Register an async wait on the backing object
3. Attempt the operation again. If it says it would block, continue.
4. Block on the waiter and then go back to (2)

Attempting the operation will effectively poll the signal state
(assuming that the wait is consistent with the operation) so we're
basically performing a regular level triggered wait with extra steps by
registering for rising edges first then polling the state.

There's still a semantic mismatch between the Linux event-driven model
and Zircon's level triggered signals which can manifest with epoll's
EPOLLET flag. The EDGE_TRIGGERED flag isn't helping us here, since we
also poll the state in the epoll logic.

In practice, most of the time the logic interested in waiting on an
object will respond by driving the level to zero. For example, code
waiting for a file to become readable will often proceed to read all
available data thus making the file no longer readable. When this
happens the next event on the file will also cause the associated signal
to be asserted and trigger a rising edge. In this situation, waiting for
an event or a level or a rising edge are all equivalent since those all
occur at the same time. It's only when the level is not driven to zero
that there is an observable difference between these modes.

Change-Id: I29624c5afb34a48157c4d43db4fbfdde56dc43db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1380076
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: James Robinson <jamesr@google.com>