[starnix] Implement Starnix-internal membarrier registration
This adds a Starnix only implementation of the membarrier registration
mechanism for "private" barriers. This tracks registrations and returns
errors in the appropriate places. It does not limit barriers that are
issued - all barriers are still global if they are issued. Registration
for "global" barriers is not supported in this change.

Bug: 297526152
Change-Id: If506ca25600d31a577d3a5483b766d378f637483
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1375961
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>