[assembly] Do not prepend "product_bundle_" in get_version tool
"ffx product get_version" provides a sanitized name for the given
artifact that can be used when communicating with the versioning server.

Platforms, products and boards are sanitized by adding the artifact type
as a prefix, but we should not do that for product bundles.

Change-Id: Ie70992daccf98baea426cfbd856e94c2f00ce92f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1384881
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>