[unwinder][EHABI] Respect pc_is_return_address
Which will cause unwinding to fail when the last instruction of a
particular function is a call instruction and sets the link register to
the first instruction of whatever function is placed after it.

Now the ArmEhabiUnwinder follows the same pattern of subtracting 1 from
the PC to get the correct function that we're unwinding from.

Fixed: 449705492
Change-Id: I6a6a696414c33f8546af784de4782b6e0e84ad52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1385662
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>