[components] Fix debug routing for void dictionaries
When `debug=true` is passed to a recursive dictionary lookup, and the
route terminates in RouterResponse::Unavailable - aka. a `void`
dictionary capability in CML - the callee propagates this response,
which breaks the contract that states the router must return
RouterResponse::Debug. Fix it to return such and populate it with a
payload of CapabilitySource::Void.

This fix is needed for scrutiny to interpret routes from dictionaries
correctly when the dictionary route comes from `void`.

Bug: 448189547
Change-Id: I621c0f474e9e8f2e116eef41b1ace3b34f143452
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1382629
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Gary Bressler <geb@google.com>