[kernel][jtrace] Fix some jtrace edge cases.
First
The `jtrace_enabled` build option is not actually a bool, it is a
tri-state which can be `true`, `false`, or `"persistent"`.  Because of
this, when evaluating `jtrace_enabled` in `gn` logic, it is important
to not use shorthand like `if (!jtrace_enabled)` in `gn` code.
Instead, we want to explicitly say something like `if (jtrace_enabled
== false)` as the boolean `!` operator only works on booleans, not
strings.  So, fix this.  This allows folks to build with persistent
jtracing enabled again.

Second
Change the definition of the JTRACE macro itself to depend on whether
or not JTRACE was enabled.  Prior to this, it was always attempting to
instantiate a JTRACE record, whose definition depended on large vs.
small record configuration (And assumed small).  So, if someone had a
bunch of large record style JTRACE statements in code, and disabled
JTRACE in their build configuration, instead of the statements simply
dropping out, they would break the build.

Instead of doing this, when jtrace is disabled, simply define the
macros to demand a `;`, but then just disappear, allowing people to
leave statements in an experimental patch stack, but also turn jtrace
off and still be be able to build.

Change-Id: Ia453064e55407feb336b4342c8822d87b526ebc4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386002
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>