[kernel][vm] Only decrement share count for PageOrRef
DecrementCowContentShareCount expects the content to either be a page or
a compressed reference. Fix a path in SupplyPagesLocked that tries to
call it with a zero marker too.

The added vmo unittest fails without this fix.

Run-All-Tests: true
Change-Id: I38d0bb23cb41bdc54899ae9ba7af3e7e00c3b09b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386001
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Sage Barreda <sagebarreda@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>