[sestarnix] Check Binder file transfers
This "greens" (flips from failing to passing) subtest seven of the
SELinux Test Suite's Binder test and this also "greens" subtests
thirteen and fifteen of the SELinux Test Suite's BPF test (and since
they were the last two subtests, this greens that test).

To avoid duplicating logic across todo_-prefixed and non-todo_-prefixed
functions, we introduce todo_option_-prefixed functions that take an
Option<BugRef> as a parameter and based on the Some-ness of the BugRef
internally call the appropriate todo_-prefixed or non-todo_-prefixed
functions of the next level down the abstraction hierarchy (and/or call
other todo_option_-prefixed functions of the next level down).

Fixes: 364569358
Bug: 398654361
Fixes: 398654362
Bug: 448598665
Change-Id: I6a9b5ebe2d87434fe0dafdd5db72f45b43f9e351
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1381484
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Reviewed-by: Ambre Williams <ambre@google.com>