[msd-vsi-vip] Disable pulse eater on reset
Disable the pulse eater and power when we go to do a reset.

This might fix the problems we are seeing where our hardware
is getting wedged and is not coming back even after a reset.

Calling `PowerOn` during the reset causes the `ResetAfterSubmit`
test to fail in debug mode because Reset is not being called on the
device thread. To fix this, have the test schedule a request to reset
on the device thread.

Bug: 439669949
Change-Id: Iad78b98c23f20158a2fe45aa9c1beb12f4349a45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1381723
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>