[starnix] Skip pending signals check if task is freezing
Cgroup freeze is implemented as a kernel signal and is handled before
other user signals. A frozen Task ignores all other signals until it is
thawed. The only exception is SIGKILL, which thaws the task before
enqueuing. If both Freeze and SIGKILL are queued on a task, avoid
freezing as it will terminate.

This CL fixes a problem where a frozen Task spins indefinitely while
trying to freeze, because it has a pending signal in its queue.

Bug: 447499791
Change-Id: I6382e34133a96efa0f67336d648cb8a9fcafec59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1384863
Fuchsia-Auto-Submit: David Song <wintermelons@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: David Song <wintermelons@google.com>