[fxtest] Allow running all repeated tests on failure
The default behavior of fx test is to stop running a repeated test when
one attempt fails. This CL adds the --[no-]fail-by-group flags to
control the behavior. When --no-fail-by-group is passed, all repeats of
a test will be attempted even if there is a failure.

Fixed: 433975024
Change-Id: Id7479b5a3265926f8bf4d4585f5ef97c9c2d1169
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386301
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>