[audio][device_registry] scheduler profile
Based on observations of slowness when establishing audio
streams if the system is highly loaded, put the ADR's thread
into a Scheduler Profile.

This uses profile "fuchsia.audio.device.registry.dispatch".

Failing to apply a Scheduler Profile is not fatal -- it may
happen in tests, or in certain products that do not include
RoleManager -- so we do not return a blocking error if this
occurs. However, we do warn because performance may suffer.

Fixed: 449772594
Bug: 442904008

Change-Id: I90803acd9ffb36c505a9b178e3a762b20acda69c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1384433
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>