[vfs][rust] Validate names in TreeBuilder
TreeBuilder wasn't checking for empty names, slashes, dots, or nulls in
names when an entry was added. If a name was invalid then adding the
entry to the Simple directory would panic.

This change validates that the names are valid when added to
TreeBuilder.

Bug: b/447366872
Change-Id: I94904fe881ac714c9ed92f9847fbc2b38918cad9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1383361
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>