[netlink][netcfg] netcfg installs rules for locally-provisioned devices
Before this change, netlink translates the `unreachable` target with
a workaround so that unmarked sockets can lookup the main table for
routes provisioned for the locally-provisioned devices. This change
shifts the responsibility from netlink to netcfg so that:

- netlink translates the rules more faithfully.
- in a product definition where locally-provisioned devices are not
required, the rule itself can go away.

Fixed: 418849362
Test: OOBE, inspected the rules
Change-Id: I6a6a6964c0616fc48d35dc641dbae4778516ad1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1360789
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>