[ffxutil] Fix processing of test failures when run with ffx test.
ffx test returns a non-zero exit code when the test fails, so we were
returning the error and not processing the output summary. We still
happened to parse out case results because the test result ended up
going through testparser because it had no outputfiles attached and
testparser was able to parse out test cases for the most part. The
noticeable difference is that case outputfiles were all attached to the
test suite result as opposed to the test case result. This change should
process test failures so that the test case artifacts appear in the
outputfiles of the test case.

Change-Id: I43a63e3f3a64e79fc657aa5e5a022a46209afa12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1387004
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Carver Forbes <carverforbes@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>