[netstack3] per-connection TCP Timestamp state
Define the structs that will hold state for the TCP Timestamp option for
each connection. At the moment, this is all dead-code. A follow up
change will start using it in the TCP state machine.

Along with laying out the state, this CL implements the logic needed to
negotiate the timestamp option with a TCP connection's peer.

Bug: 436529062
Change-Id: Ie6bf020f2bae774dc2cf9e7d3408085cc227e8d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1384522
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>