[scenic][display] Add heuristics to skip CheckConfig().
Small configuration differences can sometimes cause use to thrash the
CoordinatorProxy's CheckConfig() cache.  Based on empirical
measurements, a heuristic is added which detects situations where we
know that CheckConfig() would fail.

Now that this heuristic is in place, we can choose a better value for
the maximum cache size.  This CL bumps it up from 5 to 20, which costs
a total of ~5KB (assuming 2-layer display configs).

Bug: 446042966
Fixed: 450339197
Change-Id: I80eebfecca32606eaa9adfb39a125ca53cfeb6dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386966
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>