[sestarnix] Implement and call the perf_event_open LSM hook.
In this CL the perf_event_open hook is called only in the case when
the syscall argument for `tid` (called `pid` on linux) is 0 or -1. The
general case of `tid` > 0 is not yet handled on Starnix, and thus the
hook is not yet called for that case.

Bug: 405102319
Change-Id: I0c64a581b1719afaad720844befa5255a63a89c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1383221
Reviewed-by: Ambre Williams <ambre@google.com>
Reviewed-by: Maria Glukhova <mariagl@google.com>
Commit-Queue: Nelly Vouzoukidou <nellyv@google.com>