[starnix][task] Move ProcessGroups to their own hashmap
Rather than storing the reference to the ProcessGroup in the same hash
table as the Task objects, this CL makes a separate HashMap for
ProcessGroups. This change also makes the PidTable have a strong
reference to the ProcessGroup object because a ProcessGroup should be
removed from the PidTable before being destroyed.

Bug: b/445230228
Change-Id: I834d1f0d106fa92fbdc230b041cbc182bcbfd3cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1387781
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>