[memory_monitor2] Fix a zombie principal crash
In some cases, when `memory_monitor2` is slow processing claims, it
can happen that by the time we process a claim for a short-lived
process, the process has already exited. We used to skip processing
the principal itself, but we would keep zombie references to the
resources. This causes some issues down the line when a client
attempts to resolve attribution data, stumbles upon the zombie
resources, but references to the zombie principal cannot be resolved.

With this change, when we detect this situation, we don't skip
processing the zombie principal; we process it as normal, except we
tag it as a zombie principal, so that it can be consumed as usual for
all intents and purposes.

Bug: 445043211

Change-Id: I975044330ce9d9d92925b4b76f03f72dbe0f3ac5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1379661
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Fuchsia-Auto-Submit: Guillaume Chelfi <chelfi@google.com>
Commit-Queue: Guillaume Chelfi <chelfi@google.com>