[fdio] memfd_create uses an UNBOUNDED vmo
Create the backing VMO for memfd_create using ZX_VMO_UNBOUNDED instead
of ZX_VMO_RESIZABLE.

We are migrating userspace away from creating resizable VMOs and
attempting to increase their size with zx_stream_writev.

This is in preparation for a kernel-side change to align
zx_stream_writev to the behavior described in RFC 238, where
zx_stream_writev is said to not increase the size of the underlying VMO.

This is part of the kernel's "content size cleanup" efforts.

Run-All-Tests: true
Bug: 448723572
Change-Id: I6eae9460929eb0607be4120f32755edfdc668b16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1386098
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Ethan Williams <ethanws@google.com>