[zx][rs] stream tests operate on UNBOUNDED VMOs
zx::stream tests use ZX_VMO_UNBOUNDED instead of ZX_VMO_RESIZABLE when
attempting to add new content with zx_stream_writev.

We are migrating userspace away from creating resizable VMOs and
attempting to increase their size with zx_stream_writev.

This is in preparation for a kernel-side change to align
zx_stream_writev to the behavior described in RFC 238, where
zx_stream_writev is said to not increase the size of the underlying VMO.

This is part of the kernel's "content size cleanup" efforts.

Run-All-Tests: true
Bug: 448723572
Change-Id: I7092d1b1cfb9b2c9129fc9f34df712cf479eca6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1388021
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Ethan Williams <ethanws@google.com>