[libc] Check committed bytes for calloc'd VMO instead of committed_change_events
Depending on the allocator, the VMO backing calloc could be larger than
the actual requested calloc size. This is the case for ASan which
allocates extra space for metadata specific to ASan before the actual
allocation backing the calloc. ASan can write to this metadata which
commits pages from the VMO behind the whole calloc. It's entirely
possible but unlikely for the page scanner to reclaim these touched
metadata pages and this increments `committed_change_events` which
actually checks the number of reclamation events, not the number of
committed events. Hence, `committed_change_events` can infrequently be
non-zero.

To account for calloc'd VMOs where a small subset of the VMO is touched
by the allocator, we can instead assert that the portion of the VMO
backing the calloc is uncommitted via `populated_bytes`. The VMO can
have some bytes committed but for calloc, the number of uncommitted
bytes should be at least the requested allocation size.

Fixed: 443093612
Change-Id: Ifb22a5747ff9c55e188893e582c1775abad1d9ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1388025
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>