[component] improve `explore` tool discovery
Previously, a user of `ffx component explore` had to pass `--tools` if
they wanted additional shell commands to be available in the PATH of
the shell that `component explore` creates.

This required finding specific documentation for each component to be
explored, and users often forgot to specify tools.

To get around this, we use the `facets` field of a component manifest
which is designed for component-specific data. If a component
specifies a set of package URLs in `fuchsia.dash.launcher-tool-urls",
those package URLs will be loaded as PATH items for invocations of
`ffx component explore` on that component. If a user specifies
`--tools` on the command line, those specified on the command line
will be used instead.

Fixed: 406795126
Test: added unit tests for the new facets parsing
Change-Id: Id070f18433d5a975f55e744c0660b474e52099f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1367477
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: John Wittrock <wittrock@google.com>