[netstack3] Move data notifier out of core
Remove the DataNotifier trait from core and stop storing the data
notifier directly in core's TCP socket state; instead, move it into the
receive buffer defined by bindings, so all the notification logic
happens in bindings.

For active open sockets, the notifier comes from the
fuchsia.posix.socket/SocketCreationOptions and is stashed in the
UnconnectedSocketData to be stored in the receive buffer when the socket
is connected. For passive open sockets, the notifier is inherited from
the listening socket, and is inserted into the receive buffer by
bindings when it accepts the socket from the accept queue.

Test: netstack-power-integration-test
Change-Id: Ibffaedc3a53a63326989ee11c770f30a52b650e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1388236
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>